+++ modules/apps/portal-language/portal-language-extender-test/src/testIntegration/resources/com/liferay/portal/language/extender/internal/osgi/util/tracker/test/dependencies/batch/batch/Language_yy_ZZ.properties
