+++ modules/apps/static/portal-lpkg-deployer/portal-lpkg-deployer-impl/src/main/java/com/liferay/portal/lpkg/deployer/internal/osgi/util/tracker/WABWrapperBundleTrackerCustomizer.java
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.portal.lpkg.deployer.internal;
+package com.liferay.portal.lpkg.deployer.internal.osgi.util.tracker;
 
 import com.liferay.petra.string.StringPool;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
+import com.liferay.portal.lpkg.deployer.internal.WABWrapperUtil;
 
 import java.net.MalformedURLException;
 import java.net.URL;
